+++ source/B2B.Transactions.Api/B2BTransactionQueueTrigger.cs
@@ -17,7 +17,7 @@ using System.Text;
 using System.Threading.Tasks;
 using B2B.Transactions.Infrastructure.Configuration.Correlation;
 using B2B.Transactions.Infrastructure.Serialization;
-using B2B.Transactions.UseCases;
+using B2B.Transactions.Transactions;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
