+++ source/B2B.Transactions.Infrastructure/DataAccess/Transaction/TransactionEntityConfiguration.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using B2B.Transactions.UseCases;
+using B2B.Transactions.Transactions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
