+++ source/B2B.Transactions.Infrastructure/Transactions/TransactionQueueDispatcher.cs
@@ -19,7 +19,7 @@ using System.Transactions;
 using Azure.Messaging.ServiceBus;
 using B2B.CimMessageAdapter.Transactions;
 using B2B.Transactions.Infrastructure.Serialization;
-using B2B.Transactions.UseCases;
+using B2B.Transactions.Transactions;
 
 namespace B2B.Transactions.Infrastructure.Transactions
 {
