+++ source/B2B.Transactions/Xml/Outgoing/AcceptDocumentProvider.cs
@@ -15,7 +15,7 @@
 using System.Text;
 using System.Xml;
 using B2B.Transactions.OutgoingMessages;
-using B2B.Transactions.UseCases;
+using B2B.Transactions.Transactions;
 using Energinet.DataHub.MarketRoles.Domain.SeedWork;
 
 namespace B2B.Transactions.Xml.Outgoing
