+++ source/B2B.Transactions/Xml/Outgoing/IDocumentProvider.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using B2B.Transactions.UseCases;
+using B2B.Transactions.Transactions;
 
 namespace B2B.Transactions.Xml.Outgoing
 {
